feat: make scrollbar slider transparent (#1)
Previously, the scrollbar sliders were opaque which made all highlights/gutters in the minimap (on the right edge of the editor) behind the slider invisible. It is a common practice to make the scrollbar slider transparent. [1]

Design note: I chose the alpha value so that it looks ok on my screen. I used the old active color als hover color and linearily extrapolated the new active color from the other colors. I am not designer. If you do not find the choice of the colors perfect or if they have been copied exactly from Luna, I am open to your suggestions for improvement. :)

[1] https://sourcegraph.com/search?q=%22scrollbarSlider.background%22%3A&groupBy=repo